Cleaned up the 'reducer.per.bucket' code in BnP. - Replaced HadoopStoreBuilder with the code from HadoopStoreWriterPerBucket,   since the latter offers a superset of the former's functionality. - Similarly, replaced HadoopStoreWriterTest with the code from   HadoopStoreWriterPerBucketTest. - Deleted AvroStoreBuilderReducerPerBucket and HadoopStoreBuilderReducerPerBucket. - Modified HadoopStoreBuilder to remove unnecessary branching logic.
With these changes, BnP supports exactly the same functionality, but with a
lot less code.
